Update build system for modern bazel toolchains
- Pin bazel to 6.4.0 via .bazelversion (native cc rules + WORKSPACE
  support were removed in bazel 8/9)
- Bump googletest 1.10.0 -> 1.12.1 (old version referenced the
  removed @bazel_tools//platforms)
- Add SDK libc++ include path in .bazelrc to work around a broken
  CommandLineTools libc++ path on macOS
- Fix -Werror format warning in benchmark (nanoseconds::rep is
  long long: use %lld with a cast)

+++ benchmark/benchmark_transform.cpp
@@ -53,7 +53,7 @@ int main(int argc, char *argv[]) {
         profile_mat4(chain_mat4_f);
         auto dur =
             std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - start);
-        printf("mat4 elapsed time:   %ld ns\n", dur.count());
+        printf("mat4 elapsed time:   %lld ns\n", (long long)dur.count());
 
         // profile transf (float)
         srand(seed);
@@ -63,7 +63,7 @@ int main(int argc, char *argv[]) {
         start = std::chrono::high_resolution_clock::now();
         profile_transform(chain_transf_f);
         dur = std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - start);
-        printf("transf elapsed time: %ld ns\n", dur.count());
+        printf("transf elapsed time: %lld ns\n", (long long)dur.count());
 
         printf("[FLOAT] benchmark complete:\n");
         const auto &tm = math::translation(chain_mat4_f.back());
@@ -84,7 +84,7 @@ int main(int argc, char *argv[]) {
         profile_mat4(chain_mat4_d);
         auto dur =
             std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - start);
-        printf("mat4 elapsed time:   %ld ns\n", dur.count());
+        printf("mat4 elapsed time:   %lld ns\n", (long long)dur.count());
 
         // profile transf (double)
         srand(seed);
@@ -94,7 +94,7 @@ int main(int argc, char *argv[]) {
         start = std::chrono::high_resolution_clock::now();
         profile_transform(chain_transf_d);
         dur = std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::high_resolution_clock::now() - start);
-        printf("transf elapsed time: %ld ns\n", dur.count());
+        printf("transf elapsed time: %lld ns\n", (long long)dur.count());
 
         printf("[DOUBLE] benchmark complete:\n");
         const auto &tm = math::translation(chain_mat4_d.back());
